# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "抽奖系统后端API",
    "description": "一个支持多种抽奖模式的完整抽奖系统API，包含认证、活动管理、参与者管理、抽奖功能、系统管理等模块。",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队",
      "email": "buduan461@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "webhookAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Webhook Token"
      }
    },
    "schemas": {
...
      "Activity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "活动ID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "活动名称"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "活动描述"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "ended"],
            "description": "活动状态"
          },
          "lottery_mode": {
            "type": "string",
            "enum": ["offline", "online"],
            "description": "抽奖模式"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "description": "开始时间"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "description": "结束时间"
          },
          "settings": {
            "type": "object",
            "properties": {
              "max_lottery_codes": {
                "type": "integer",
                "minimum": 1,
                "description": "最大抽奖码数量"
              },
              "lottery_code_format": {
                "type": "string",
                "enum": ["4_digit_number", "8_digit_number", "8_digit_alphanumeric", "12_digit_number", "12_digit_alphanumeric"],
                "description": "抽奖码格式"
              },
              "allow_duplicate_phone": {
                "type": "boolean",
                "description": "是否允许重复手机号"
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "Prize": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "奖品ID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "奖品名称"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "奖品描述"
          },
          "total_quantity": {
            "type": "integer",
            "minimum": 0,
            "description": "奖品总数量"
          },
          "remaining_quantity": {
            "type": "integer",
            "minimum": 0,
            "description": "剩余数量"
          },
          "probability": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "中奖概率"
          },
          "sort_order": {
            "type": "integer",
            "minimum": 0,
            "description": "排序值"
          }
        }
      },
      "LotteryCode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "抽奖码ID"
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "description": "抽奖码"
          },
          "status": {
            "type": "string",
            "enum": ["unused", "used"],
            "description": "使用状态"
          },
          "participant_info": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100,
                "description": "参与者姓名"
              },
              "phone": {
                "type": "string",
                "description": "手机号"
              },
              "email": {
                "type": "string",
                "format": "email",
                "description": "邮箱"
              }
            }
          },
          "used_at": {
            "type": "string",
            "format": "date-time",
            "description": "使用时间"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "LotteryRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "抽奖记录ID"
          },
          "is_winner": {
            "type": "boolean",
            "description": "是否中奖"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "抽奖时间"
          },
          "lottery_code": {
            "$ref": "#/components/schemas/LotteryCode"
          },
          "prize": {
            "$ref": "#/components/schemas/Prize"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "请求是否成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          }
        },
        "required": ["success"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "错误代码"
              },
              "message": {
                "type": "string",
                "description": "错误描述"
              },
              "details": {
                "type": "string",
                "description": "详细错误信息"
              }
            }
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "未认证",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "权限不足",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/lottery/activities/{id}": {
      "get": {
        "tags": ["抽奖模块"],
        "summary": "获取活动抽奖信息",
        "description": "获取活动抽奖信息（公开接口）",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "activity": {
                              "$ref": "#/components/schemas/Activity"
                            },
                            "prizes": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Prize"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/lottery/activities/{id}/draw": {
      "post": {
        "tags": ["抽奖模块"],
        "summary": "用户使用抽奖码参与线上抽奖",
        "description": "用户使用抽奖码参与线上抽奖",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lottery_code": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "description": "抽奖码"
                  },
                  "participant_info": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 100,
                        "description": "参与者姓名"
                      },
                      "phone": {
                        "type": "string",
                        "description": "手机号"
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "description": "邮箱"
                      }
                    },
                    "required": ["name", "phone"]
                  }
                },
                "required": ["lottery_code", "participant_info"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "抽奖成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "lottery_record": {
                              "$ref": "#/components/schemas/LotteryRecord"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/lottery/activities/{id}/offline-draw": {
      "post": {
        "tags": ["抽奖模块"],
        "summary": "管理员使用抽奖码进行线下抽奖",
        "description": "管理员使用抽奖码进行线下抽奖",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "活动ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "lottery_code": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "description": "抽奖码"
                  }
                },
                "required": ["lottery_code"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "抽奖成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "lottery_record": {
                              "$ref": "#/components/schemas/LotteryRecord"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
...
}